requirements.txt
Django==3.1
Pillow==8.0
psycopg2-binary==2.8.5
djangorestframework==3.11.0
python-decouple
drf-yasg